# swagger.yaml
openapi: 3.0.3
info:
  title: LoanStreet Backend API
  description: RESTful API for managing loan records. Supports creating, retrieving, and updating loans.
  version: 1.0.0

servers:
  - url: https://loanstreet-backend-493662086412.us-central1.run.app
    description: Production (Cloud Run)
  - url: http://localhost:8080
    description: Local development server

tags:
  - name: Loans
    description: Loan management operations

paths:
  /api/loans:
    post:
      tags:
        - Loans
      summary: Create a new loan
      description: Accepts loan properties and creates a new loan record. Returns the persisted entity with a generated UUID.
      operationId: createLoan
      parameters:
        - $ref: '#/components/parameters/XRequestID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoanCreateRequest'
      responses:
        '201':
          description: Loan created successfully
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Loan'
        '400':
          description: Invalid request payload
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Internal server error
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  /api/loans/{id}:
    get:
      tags:
        - Loans
      summary: Retrieve a loan by ID
      description: Fetches a loan record by its UUID. Returns 404 if the loan does not exist.
      operationId: getLoan
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the loan to retrieve
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/XRequestID'
      responses:
        '200':
          description: Loan found and returned
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Loan'
        '400':
          description: Invalid request — malformed UUID
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Loan not found
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Internal server error
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

    put:
      tags:
        - Loans
      summary: Update an existing loan
      description: Updates loan properties by UUID.
      operationId: updateLoan
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the loan to update
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/XRequestID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoanUpdateRequest'
      responses:
        '200':
          description: Loan updated successfully
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Loan'
        '404':
          description: Loan not found
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '400':
          description: Invalid request payload
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Internal server error
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

components:
  schemas:
    Loan:
      type: object
      description: Complete loan entity including generated ID
      properties:
        id:
          type: string
          format: uuid
          description: Auto-generated UUID for the loan
          example: "550e8400-e29b-41d4-a716-446655440000"
        amount:
          type: number
          format: decimal
          description: Loan amount
          example: 10000.00
        interestRate:
          type: number
          format: decimal
          description: Annual interest rate as a decimal
          example: 0.05
        lengthInMonths:
          type: integer
          description: Loan term in months
          example: 36
        monthlyPaymentAmount:
          type: number
          format: decimal
          description: Calculated monthly payment amount
          example: 299.71
      required:
        - id
        - amount
        - interestRate
        - lengthInMonths
        - monthlyPaymentAmount

    LoanCreateRequest:
      type: object
      description: Request payload for creating a new loan (ID is auto-generated)
      properties:
        amount:
          type: number
          format: decimal
          description: Loan amount
          example: 10000.00
          minimum: 0
          exclusiveMinimum: true
        interestRate:
          type: number
          format: decimal
          description: Annual interest rate as a decimal (0 to 1)
          example: 0.05
          minimum: 0
          exclusiveMinimum: true
          maximum: 1.0
        lengthInMonths:
          type: integer
          description: Loan term in months
          example: 36
          minimum: 1
          maximum: 600
        monthlyPaymentAmount:
          type: number
          format: decimal
          description: Calculated monthly payment amount
          example: 299.71
          minimum: 0
          exclusiveMinimum: true
      required:
        - amount
        - interestRate
        - lengthInMonths
        - monthlyPaymentAmount

    LoanUpdateRequest:
      type: object
      description: Request payload for a full replacement update of an existing loan. All fields are required.
      properties:
        amount:
          type: number
          format: decimal
          description: Loan amount
          example: 15000.00
          minimum: 0
          exclusiveMinimum: true
        interestRate:
          type: number
          format: decimal
          description: Annual interest rate as a decimal (0 to 1)
          example: 0.045
          minimum: 0
          exclusiveMinimum: true
          maximum: 1.0
        lengthInMonths:
          type: integer
          description: Loan term in months
          example: 48
          minimum: 1
          maximum: 600
        monthlyPaymentAmount:
          type: number
          format: decimal
          description: Calculated monthly payment amount
          example: 342.05
          minimum: 0
          exclusiveMinimum: true
      required:
        - amount
        - interestRate
        - lengthInMonths
        - monthlyPaymentAmount

    ApiError:
      type: object
      description: Standard error response returned by the API
      properties:
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp of when the error occurred
          example: "2026-02-21T22:22:04Z"
        error:
          type: string
          description: Short error description
          example: "Validation failed"
        errors:
          type: array
          description: List of field-level validation errors (only present for 400 validation errors)
          items:
            $ref: '#/components/schemas/FieldViolation'
      required:
        - timestamp
        - error

    FieldViolation:
      type: object
      description: Individual field validation error
      properties:
        field:
          type: string
          description: Name of the field that failed validation
          example: "amount"
        message:
          type: string
          description: Human-readable error message
          example: "must be greater than 0"
      required:
        - field
        - message

  parameters:
    XRequestID:
      name: X-Request-ID
      in: header
      required: false
      description: Unique request identifier for request tracing. If not provided, a new one will be generated.
      schema:
        type: string
        format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"

  headers:
    XRequestID:
      required: true
      description: Unique request identifier for request tracing. Always present — echoed from request or server-generated if not provided.
      schema:
        type: string
        format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"